when deciding which peers to try to connect to, add some weights to punish peers that don't transfer piece data with us.  We'll still give them another, but at longer intervals than better peers get.

+++ libtransmission/peer-mgr.h
@@ -43,13 +43,13 @@ void tr_peerMgrAddIncoming( tr_peerMgr      * manager,
 
 void tr_peerMgrAddPeers( tr_peerMgr     * manager,
                          const uint8_t  * torrentHash,
-                         int              from,
+                         uint8_t          from,
                          const uint8_t  * peerCompact,
                          int              peerCount );
 
 void tr_peerMgrAddPex( tr_peerMgr     * manager,
                        const uint8_t  * torrentHash,
-                       int              from,
+                       uint8_t          from,
                        const tr_pex   * pex,
                        int              pexCount );
